+++ Enhetsregisteret/AT.Common.Enhetsregisteret.Publish/Implementation/Constants.cs
@@ -6,5 +6,4 @@ namespace Arbeidstilsynet.Common.Enhetsregisteret.Implementation;
 internal static class Constants
 {
     public const int MaxSearchResultSize = 10_000;
-    public const int MaxSearchStringLength = 180;
 }